feat: move store logging to separate package
There is now the option to attach "stale subscribers" to the store which are
notified when an action are applied but the state hasn't been changed.

There is now an extra option to apply changes to the store when it is added
to a plugin's context.

BREAKING CHANGE: Store logging must be done via subscribers